micro: extract composer attachment state
