+++ frontend/src/app/agent/_components/use-composer-attachments.ts
@@ -1,0 +1,118 @@
+"use client";
+
+import { useCallback, useState, type DragEvent, type RefObject } from "react";
+import type { SessionTab } from "@/lib/agent/session";
+import {
+  attachmentDedupKey,
+  createAttachment,
+  dataTransferHasFiles,
+  filesFromDataTransfer,
+  type ChatAttachment,
+} from "./chat-attachments";
+
+type UpdateTab = (tabId: string, patch: (tab: SessionTab) => SessionTab) => void;
+
+type UseComposerAttachmentsOptions = {
+  activeTab: SessionTab | null;
+  running: boolean;
+  updateTab: UpdateTab;
+  fileInputRef: RefObject<HTMLInputElement | null>;
+};
+
+export function useComposerAttachments({
+  activeTab,
+  running,
+  updateTab,
+  fileInputRef,
+}: UseComposerAttachmentsOptions) {
+  const [attachments, setAttachments] = useState<ChatAttachment[]>([]);
+  const [readingAttachments, setReadingAttachments] = useState(false);
+  const [composerDragActive, setComposerDragActive] = useState(false);
+
+  const attachFiles = useCallback(
+    async (files: FileList | File[] | null) => {
+      const fileArray = files ? Array.from(files) : [];
+      if (fileArray.length === 0 || !activeTab) return;
+      if (running) {
+        updateTab(activeTab.id, (tab) => ({
+          ...tab,
+          error: "Pause or wait for the current turn before attaching files.",
+        }));
+        return;
+      }
+      setReadingAttachments(true);
+      try {
+        const next = await Promise.all(fileArray.map((file) => createAttachment(file)));
+        setAttachments((current) => {
+          const seen = new Set(current.map(attachmentDedupKey));
+          const uniqueNext: ChatAttachment[] = [];
+          next.forEach((file) => {
+            const key = attachmentDedupKey(file);
+            if (seen.has(key)) return;
+            seen.add(key);
+            uniqueNext.push(file);
+          });
+          return [...current, ...uniqueNext];
+        });
+        updateTab(activeTab.id, (tab) => ({ ...tab, error: "" }));
+      } catch (err) {
+        updateTab(activeTab.id, (tab) => ({
+          ...tab,
+          error: err instanceof Error ? err.message : "Failed to attach file",
+        }));
+      } finally {
+        setReadingAttachments(false);
+        if (fileInputRef.current) fileInputRef.current.value = "";
+      }
+    },
+    [activeTab, fileInputRef, running, updateTab],
+  );
+
+  const removeAttachment = useCallback((id: string) => {
+    setAttachments((current) => current.filter((item) => item.id !== id));
+  }, []);
+
+  const clearAttachments = useCallback(() => {
+    setAttachments([]);
+    if (fileInputRef.current) fileInputRef.current.value = "";
+  }, [fileInputRef]);
+
+  const handleComposerDragOver = useCallback(
+    (event: DragEvent<HTMLDivElement>) => {
+      if (!dataTransferHasFiles(event.dataTransfer)) return;
+      event.preventDefault();
+      event.dataTransfer.dropEffect = running ? "none" : "copy";
+      setComposerDragActive(true);
+    },
+    [running],
+  );
+
+  const handleComposerDragLeave = useCallback((event: DragEvent<HTMLDivElement>) => {
+    const nextTarget = event.relatedTarget;
+    if (nextTarget instanceof Node && event.currentTarget.contains(nextTarget)) return;
+    setComposerDragActive(false);
+  }, []);
+
+  const handleComposerDrop = useCallback(
+    (event: DragEvent<HTMLDivElement>) => {
+      if (!dataTransferHasFiles(event.dataTransfer)) return;
+      event.preventDefault();
+      setComposerDragActive(false);
+      void attachFiles(filesFromDataTransfer(event.dataTransfer));
+    },
+    [attachFiles],
+  );
+
+  return {
+    attachments,
+    setAttachments,
+    readingAttachments,
+    composerDragActive,
+    attachFiles,
+    removeAttachment,
+    clearAttachments,
+    handleComposerDragOver,
+    handleComposerDragLeave,
+    handleComposerDrop,
+  };
+}
